+++ lib/spellr/check_dry_run.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+require_relative '../spellr'
+require_relative 'check'
+
+module Spellr
+  class CheckDryRun < Check
+    def check
+      files.each do |file|
+        puts file.relative_path
+      end
+    end
+  end
+end