+++ lib/spellr/cli.rb
@@ -1,10 +1,8 @@
 # frozen_string_literal: true
 
-require 'optparse'
-require 'pathname'
-require 'open3'
-
 require_relative '../spellr'
+require_relative 'cli_options'
+require_relative 'string_format'
 
 module Spellr
   class CLI
@@ -13,15 +11,22 @@ module Spellr
     def initialize(argv)
       @argv = argv
 
-      parse_command
+      CLI::Options.parse(@argv)
+
+      run
+    end
+
+    def run
+      validate_config
+
+      check
+    rescue Spellr::Error => e
+      warn red("Error: #{e.message}")
+      exit 1
     end
 
     def check
-      require_relative 'check'
-
-      validate_config
-
-      checker = Spellr::Check.new(files: files)
+      checker = Spellr.config.checker.new(files: files)
       checker.check
 
       exit checker.exit_code
@@ -36,78 +41,7 @@ module Spellr
 
     def files
       require_relative 'file_list'
-      Spellr::FileList.new(*argv)
-    end
-
-    def wordlist_option(_)
-      require_relative 'wordlist_reporter'
-      Spellr.config.reporter = Spellr::WordlistReporter.new
-    end
-
-    def quiet_option(_)
-      Spellr.config.quiet = true
-      require_relative 'quiet_reporter'
-      Spellr.config.reporter = Spellr::QuietReporter.new
-    end
-
-    def interactive_option(_)
-      require_relative 'interactive'
-      Spellr.config.reporter = Spellr::Interactive.new
-    end
-
-    def config_option(file)
-      Spellr.config.config_file = Pathname.pwd.join(file).expand_path
-    end
-
-    def dry_run_option(_)
-      files.each { |f| puts f.relative_path_from(Pathname.pwd) }
-
-      exit
-    end
-
-    def version_option(_)
-      require_relative 'version'
-      puts(Spellr::VERSION)
-
-      exit
-    end
-
-    def parse_command
-      parse_options
-      check
-    end
-
-    def options_help(_)
-      puts options.help
-
-      exit
-    end
-
-    def parse_options
-      options.parse!(argv)
-    end
-
-    def options # rubocop:disable Metrics/MethodLength, Metrics/AbcSize
-      @options ||= begin
-        opts = OptionParser.new
-
-        opts.banner = 'Usage: spellr [options] [files]'
-        opts.separator('')
-        opts.on('-w', '--wordlist', 'Outputs errors in wordlist format', &method(:wordlist_option))
-        opts.on('-q', '--quiet', 'Silences output', &method(:quiet_option))
-        opts.on('-i', '--interactive', 'Runs the spell check interactively',
-                &method(:interactive_option))
-        opts.separator('')
-        opts.on('-d', '--dry-run', 'List files to be checked', &method(:dry_run_option))
-        opts.separator('')
-        opts.on('-c', '--config FILENAME', String, <<~HELP, &method(:config_option))
-          Path to the config file (default ./.spellr.yml)
-        HELP
-        opts.on('-v', '--version', 'Returns the current version', &method(:version_option))
-        opts.on('-h', '--help', 'Shows this message', &method(:options_help))
-
-        opts
-      end
+      Spellr::FileList.new(argv)
     end
   end
 end