+++ lib/spellr/file_list.rb
@@ -8,17 +8,13 @@ module Spellr
   class FileList
     include Enumerable
 
-    def initialize(*patterns)
+    def initialize(patterns = nil)
       @patterns = patterns
     end
 
     def each
       fast_ignore.each do |file|
-        next unless cli_patterns_ignore.allowed?(file)
-
-        file = Spellr::File.new(file)
-
-        yield(file)
+        yield(Spellr::File.new(file))
       end
     end
 
@@ -28,28 +24,12 @@ module Spellr
 
     private
 
-    # anchored patterns are significantly faster on large codebases
-    def cli_patterns
-      @patterns.map do |pattern|
-        pattern.sub(%r{^(?!(?:[/~*]|\.{1,2}/))}, '/')
-      end
-    end
-
-    def gitignore_path
-      gitignore = ::File.join(Dir.pwd, '.gitignore')
-      gitignore if ::File.exist?(gitignore)
-    end
-
     def fast_ignore
       FastIgnore.new(
         ignore_rules: Spellr.config.excludes,
         include_rules: Spellr.config.includes,
-        gitignore: gitignore_path
+        argv_rules: @patterns
       )
-    end
-
-    def cli_patterns_ignore
-      @cli_patterns_ignore ||= FastIgnore.new(include_rules: cli_patterns, gitignore: false)
     end
   end
 end