+++ lib/spellr/key_tuner/possible_key.rb
@@ -2,6 +2,7 @@
 
 require 'pathname'
 require_relative 'stats'
+require_relative '../backports'
 
 class PossibleKey # rubocop:disable Metrics/ClassLength
   include Stats